// src/protos/scroller_pb.ts
// @generated by protoc-gen-es v2.12.1 with parameter "target=ts"
// @generated from file scroller.proto (syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file scroller.proto.
 */
export const file_scroller: GenFile = /*@__PURE__*/
  fileDesc("Cg5zY3JvbGxlci5wcm90byJQCgxTY3JvbGxlckRhdGESDwoCaWQYASABKANIAYgBARIOCgRwYWdlGAIgASgNSAASEAoGdXBkYXRlGAMgASgISABCBgoEZGF0YUIFCgNfaWQiLwoQQmFzZVNjcm9sbGVyRGF0YRIbCgRkYXRhGAEgASgLMg0uU2Nyb2xsZXJEYXRhYgZwcm90bzM");

/**
 * @generated from message ScrollerData
 */
export type ScrollerData = Message<"ScrollerData"> & {
  /**
   * user's id
   *
   * @generated from field: optional int64 id = 1;
   */
  id?: bigint | undefined;

  /**
   * @generated from oneof ScrollerData.data
   */
  data: {
    /**
     * new page
     *
     * @generated from field: uint32 page = 2;
     */
    value: number;
    case: "page";
  } | {
    /**
     * update cache
     *
     * @generated from field: bool update = 3;
     */
    value: boolean;
    case: "update";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message ScrollerData.
 * Use `create(ScrollerDataSchema)` to create a new message.
 */
export const ScrollerDataSchema: GenMessage<ScrollerData> = /*@__PURE__*/
  messageDesc(file_scroller, 0);

/**
 * @generated from message BaseScrollerData
 */
export type BaseScrollerData = Message<"BaseScrollerData"> & {
  /**
   * user's id and page or update
   *
   * @generated from field: ScrollerData data = 1;
   */
  data?: ScrollerData | undefined;
};

/**
 * Describes the message BaseScrollerData.
 * Use `create(BaseScrollerDataSchema)` to create a new message.
 */
export const BaseScrollerDataSchema: GenMessage<BaseScrollerData> = /*@__PURE__*/
  messageDesc(file_scroller, 1);